+++ src/trio_util/_async_value.py
@@ -1,5 +1,5 @@
 from collections import namedtuple
-from contextlib import asynccontextmanager, contextmanager
+from contextlib import contextmanager, ExitStack
 from functools import partial
 
 import trio
@@ -270,9 +270,9 @@ class AsyncValue:
             yield output
 
 
-@asynccontextmanager
-async def compose_values(**value_map):
-    """Async context manager providing a composite of multiple AsyncValues
+@contextmanager
+def compose_values(**value_map):
+    """Context manager providing a composite of multiple AsyncValues
 
     The composite object itself is an AsyncValue, with the `value` of each
     underlying object accessible as attributes on the composite `value`.
@@ -288,39 +288,29 @@ async def compose_values(**value_map):
     Synopsis:
         >>> async_x, async_y = AsyncValue(-1), AsyncValue(10)
         >>>
-        >>> async with compose_values(x=async_x, y=async_y) as async_xy:
+        >>> with compose_values(x=async_x, y=async_y) as async_xy:
         >>>     result = await async_xy.wait_value(lambda val: val.x < 0 < val.y))
         >>>
         >>> result
         CompositeValue(x=-1, y=10)
     """
-    async with trio.open_nursery() as nursery:
-        values = value_map.values()
-        # the implementation requires `wait_transition()` taking a predicate
-        if not (values and all(isinstance(e, AsyncValue) for e in values)):
+    with ExitStack() as stack:
+        async_vals = value_map.values()
+        if not (async_vals and all(isinstance(av, AsyncValue) for av in async_vals)):
             raise TypeError('expected instances of AsyncValue')
         value_type = namedtuple('CompositeValue', value_map.keys())
-        # NOTE: initial None values are replaced before yielding
-        composite = AsyncValue(value_type._make((None, ) * len(values)))
-        listener_count = AsyncValue(0)
+        composite = AsyncValue(value_type._make(av.value for av in async_vals))
 
-        # This predicate is run indefinitely per underlying value and updates
+        # This dummy wait_value() predicate hooks into each value and updates
         # the composite as a side effect.
-        def _update_composite(name, val, _):
+        def _update_composite(name, val):
             composite.value = composite.value._replace(**{name: val})
             return False
 
-        async def _wait_transition(name, e):
-            listener_count.value += 1
-            await e.wait_transition(partial(_update_composite, name))
-
-        for name_, e in value_map.items():
-            nursery.start_soon(_wait_transition, name_, e)
-
-        # Before yielding, ensure the listen functions are blocked,
-        # and that the composed value is consistent with this blocked state.
-        await listener_count.wait_value(len(values))
-        composite.value = value_type._make(e.value for e in values)
+        for name_, async_val in value_map.items():
+            # NOTE: by using AsyncValue internals we avoid running wait_value()
+            # as a child task for each event.
+            stack.enter_context(
+                async_val._level_results.open_ref(partial(_update_composite, name_)))
 
         yield composite
-        nursery.cancel_scope.cancel()